LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing/html5lib_tests21-expected.txt

html5lib Parser Test


PASS html5lib_tests21.html 49994884ec96a706a3c1b702049a1612e43a81b5
PASS html5lib_tests21.html 9128ce0d001f8a331dadfd0091dc0faaadcc27a3
PASS html5lib_tests21.html 250daec0d65f2528955e24fbe667cf40adaee33f
PASS html5lib_tests21.html 7cdfc1d2d87a4f9ebd754d18dc7b55f73f5ba58b
PASS html5lib_tests21.html 3d12442cdaef117f92af7ca90e7e03550e07314a
PASS html5lib_tests21.html afeea6e5607448d59d16db2c2dad6e0f4f571f25
PASS html5lib_tests21.html 30b6ff244d2cd3716c1ad482e30ff294dcfed2b7
PASS html5lib_tests21.html 49b8f552d24bba176c920b40208e1d9f3d866075
PASS html5lib_tests21.html 60293a1ec80139eb1c9f05e48ca631e812a708da
PASS html5lib_tests21.html 753ad0dcb4fd3b1d6da2c87dd59116287e08d412
FAIL html5lib_tests21.html 80607dd011814b8d3ef5c9ca380fec044dd5e1aa assert_equals: expected "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <svg svg>\n|       \"foo]\"" but got "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <svg svg>\n|       \"foo]]]>\""
PASS html5lib_tests21.html b313be34a12fd540ab959b6cf18610fba03cb63f
FAIL html5lib_tests21.html c9f579bf49de2d4c553d03e43772c0d94be474c0 assert_equals: expected "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <svg svg>\n|       \"foo]]]\"" but got "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <svg svg>\n|       \"foo]]]]]>\""
PASS html5lib_tests21.html e714d65a37389755cd7385275f14a757c36b9510
PASS html5lib_tests21.html df235b5f7ba9e6e1032bae6c75e92e6a68f48cda
PASS html5lib_tests21.html fec253c6085a518dd72fdaff371da31e0d6bdb96
PASS html5lib_tests21.html 64150a9d88b6214a4a83cf9335e0b0a35e3be170
PASS html5lib_tests21.html 8c1c2ca15e12599448980201683ece12d4923d9d
PASS html5lib_tests21.html 32780672eec230ac59d4b61a1adb209bd6a0ed46
PASS html5lib_tests21.html be6b00c6bd7e576953b0bceefd9c38d14b452b8d
PASS html5lib_tests21.html 9df59cd349097dca330fefef6c1b1bbbfddadae6
PASS html5lib_tests21.html 671f606f5cb7033854d99b96b040994f0f451496
PASS html5lib_tests21.html caa80af33c7880aaddd17824efff1774ece01325
80607dd011814b8d3ef5c9ca380fec044dd5e1aa

Input

<!DOCTYPE html><svg><![CDATA[foo]]]>
Expected

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]"
Actual

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]]]>"
c9f579bf49de2d4c553d03e43772c0d94be474c0

Input

<!DOCTYPE html><svg><![CDATA[foo]]]]]>
Expected

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]]]"
Actual

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo]]]]]>"
